requirements.txt

djongo==1.3.6
django==3.2
djangorestframework<3.15
djangorestframework-simplejwt<5.3
django-cors-headers<4.2
djoser<2.3
drf_spectacular<0.27
drf-ujson2<1.8
environs<9.6
reportlab<4.1
pylint<2.14
pytest<7.3
pytest-django
# pymongo[snappy,gssapi,srv,tls]
mongoengine==0.25